Enable image caching in browser
To enable image caching in the broswer, the Cache-Control
metadata for each image in the storage must be set. This is
returned as the Cache-Control HTTP header when the browser
gets an image.

Set the metadata to "private,max-age=3600" because the
majority of the storage bandwidth is used to download:
* The user uploaded image in the open phase
* Immutable images in the voting and closed phase

In either case, the images can only be changed by the user
or they will never change. In either case, it is fine to
just let the browser cache the images for a long time like
1 hour i.e. 3600 seconds.